.
Initial open-source commit of the OpenVMM project.

In the ~5 years prior to being open-sourced, countless Microsoft
employees - both full time, and interns - helped make OpenVMM what it is
today.

While we can't share the entire commit history of the project, the
following are a few notable statistics from OpenVMM's time as a
closed-source project within Microsoft:

- initial commit:     Thu Jun 20, 2019
- #commits on `main`: 5249
- #contributors:      119