openhcl_boot: Pass the value for the DMA cache coherency (#46)
The Linux kernel does not assume that the DMA is cache coherent on arm64
whereas VMBus transation always are. No spec saying that could be found,
yet the vmbus driver in the Linux kernel makes sure that is true in the
ACPI case by checking on the _CCA property.

Here we have DT and should let the kernel DMA subsystem know that DMA on
the bus is cache coherent to save on teh CPU cache maintenance. The
kernel has all it needs for picking up this standard property from DT.

This might land some perf gains and/or uncover subtle bugs on arm64 yet
we have to have this.

For validation, I booted the guest up, and checked that the flag was set
in the driver:

```
[0.008750] hv_vmbus: >>>>> of_property_read_bool(np, "dma-coherent") == 1
[0.008750] hv_vmbus: >>>>> pdev->dev.dma_coherent == 1
```