Rename env vars to match project names (#30)
Rename `HVLITE_` and `UNDERHILL_`-prefixed environment variables to have
`OPENVMM_` and `OPENHCL_` prefixes.

In most cases, support the old names for compatibility. Update all
callers to use the new names, though; this should be safe since the
callers are always shipped together with the callees. (There is an
exception that petri will no longer be able to configure logging levels
for downlevel openhcl/underhill binaries for servicing tests. We can fix
this if it ever becomes a problem.)

Leave the documentation updates to a separate change to avoid
conflicting with ongoing guide updates.

Also remaining is the `UNDERHILL_CRASH_` prefix. I figure we can rename
this once we decide what to name the associated binary (which is
currently `underhill-crash`).