Construct and wire up lapics for VTL 1 (#17)
Make the lapics for SNP and Hv backings into an array of 2 instead of
just the 1 per-vp, and wire up the VTLs getting passed around. Next up
is likely taking the result of polling these two apics and using it to
determine which VTL to run in our next run_vp call.