Fix CI issues due to disconnected 1ES pools and unneeded AzureLogin tasks (#118)
This has been a lot of fun to figure out...

When we marked the repo as public the 1ES hosted pools that we were
previously using lost their linkage to this repository. This meant that
PR tasks were not being picked up. The fix for this was to re-create new
pools that are scoped to our repo.

The other issue this change resolves is removing the Azure Login task
from stages that do not need it. There are still some tasks that do need
access to the key vault (to download test VHD's) but this will get the
CI in a much better state.

---------

Co-authored-by: Ben Hillis <benhill@ntdev.microsoft.com>