CI: Introduce new Azure DevOps pipeline to queue private pull request validation (#148)
This change revisits the approach to queue our internal ADO PR
validation pipeline. Instead of using a GitHub workflow which uses a
python script to queue the build, this uses an ADO pipeline which uses
the TriggerBuild@4 task. After this is merged I will set up the pull
request trigger (which must be done in ADO) and validate this is working
as expected before adding it as a required check for pull requests.

---------

Co-authored-by: Ben Hillis <benhill@ntdev.microsoft.com>