loader: fix sidecar offset (#126)
If the sidecar kernel is not already 2MB aligned, then the offset is
miscomputed due to variable shadowing. Fix this.

In the future, we should get rid of this 2MB requirement, since sidecar
can use 4KB pages relatively easily and has a small enough working set
that the page size will not affect TLB pressure.

Also, fix some error strings to be more concise and useful.