virt_mshv_vtl: use GuestVtl instead of Vtl (#167)
When we know statically that we're referring to VTL0/VTL1 and not VTL2,
use the new enum `GuestVtl` instead of `Vtl`. This allows the compiler
to generate better code, it allows the reader to better understand the
invariants, and it provides more safeguards that the guest is not able
to manipulate VTL2 state.

As part of this, fix a few bugs.