flowey/gh: ensure the "all good" job does what its supposed to do (#176)
It seems that on github, this job never actually did what it was
supposed to do, which was:

- pass if all previous jobs passed
- fail if any previous job failed

Instead, it seems that in the case of any previous job failing, GitHub
Actions would simply skip executing the "all good" job entirely, and
treat the skip as a "pass" of the (required) check. An interesting
default indeed...

We never noticed this as we would previously run the PrivateVsoBuild
job, which was a superset of all open-source checks, and would therefore
fail if any open-source check failed, causing the PR to be blocked even
through the "all good" check was ostensibly "passed".

This PR should fix things up such that our "all good" job does what its
supposed to do, by relying on a workaround suggested in
https://github.com/actions/runner/issues/2566 (albeit, a slightly
tweaked variant of it).

* * * 

I also switched the job to run on a Linux machine, which I believe is
marginally faster to provision and execute on.