openhcl_boot: tell Linux not to probe for a PCI bus (#184)
By default on x86, Linux probes for a PCI bus via the usual 0xcf8
method. For OpenHCL, this is unnecessary--there is no PCI bus there with
a Hyper-V host, and for non-Hyper-V hosts we would expect the PCI bus to
be explicitly enumerated.

This reduces exits to the host during boot and it removes another
possible CVM attack vector.