ci: publish generated rustdocs to openvmm.dev (#201)
We are already building rustdocs for windows and linux in CI.
This PR adds the necessary wiring to publish them on openvmm.dev.

As part of this change, I've also moved the openvmm Guide from the root
of `openvmm.dev/`, to a `openvmm.dev/guide` subdirectory. Any attempts
to access `openvmm.dev` directly will redirect to the Guide. This will
give us the flexibility to add a shiny "landing page" to openvmm.dev at
some point in the future, without breaking any external links other
websites may have to our website.