openhcl_boot: verify hv VP index matches dt CPU index (#190)
Code in various places assumes that the hypervisor VP index is the same
as the devicetree CPU index. Validate this early, in the boot loader,
and then remove code elsewhere that half-heartedly tries to support
alternative mappings. The host already guarantees this in practice, so
let's take an explicit dependency.

If we ever weaken this requirement, we can just update the boot loader
to reorder the CPU nodes to match the actual configuration before
passing them on.