Updates to the getting started sections of the openvmm developer guide (#216)
This is a collection of several small changes to the OpenVMM developer
guide. I am setting up a new clone of this environment and following the
instructions in linear order. What prompted me to first make this change
is that I encountered an error running `cargo xflowey restore-packages`:
it seemed to need a linker. I found going "out of order" a bit jarring.
Since the same core dependencies are required to build OpenVMM and
OpenHCL (I think), and we expect contributors to set up their
environment first, I moved this up. And, I noticed a broken link.

* Fix broken link to the contributing guide.
* Move the Linux compiler / linker prereqs to the Linux / WSL2 setup
page. These are necessary before the xflowey restore.
* Use the fancy new admonish plugin for the cross compile tip.