loader: simplify handling of multiple VTLs (#217)
The individual firmware loaders generally do not need to be VTL aware.
Remove VTL awareness from most of the code. This will soon allow OpenHCL
to be loaded into VTL0.

The remaining exception is the VTL0 VP context. Simplify the handling of
this, so that the CVM-specific code doesn't have to know about this.

A future change should consider whether we really need this at all: it
seems that it's just so that OpenVMM in OpenHCL doesn't have to replay
the UEFI register state. But OpenVMM already knows so much about how to
load UEFI, so it seems kind of silly to pretend that it doesn't know how
to set the register state appropriately.