virt_mshv_vtl: Move hv and untrusted_synic into backings (#252)
This removes two more options from UhProcessor, replacing them with
getters on Backing. This results in the removal of unwraps in
backing-specific code. While it also results in slightly wordier
backing-agnostic code, it should optimize better, as each backing can
now specialize on whether certain variables are guaranteed to be present
or not.

This does not move the equivalent state from UhPartition to
BackingShared types, as since UhPartition is not generic I'd expect
there to be no meaningful optimization gains from doing so. We could
still do so for the sake of organization and self-documentation if we
wanted to though.