underhill_mem: register memory with proper private/shared bit (#257)
At some point, we stopped passing the shared bit down when registering
kernel memory. Fix this.

This fixes kernel-mode DMA from TDX VMs.

Note that there are still problems with the overall scenario: the kernel
will double buffer through the swiotlb even when the target pages are
already shared, and swiotlb is sized by openhcl_boot such that there's
not enough memory for all the IO that we may have in flight
concurrently.