underhill_crash: Trace less during crash reporting (#261)
One of our core crash diagnostics is that we send a chunk of kmsg output
to the host to log. It is ideal if that output contains the panic
message that caused the crash, but that isn't guaranteed. Since we have
a fixed size buffer, it is possible that output that occurs after the
panic message could push the useful details off the end of the buffer.
This PR reduces the amount of tracing we emit during the crash dump
process to make this less likely.