openhcl: add magic field to measured header in loader info and validate (#265)
Add a magic header to measured information that should be loaded by the
host. Additionally, zero out the VTL2 boot data range, to validate that
on servicing operations, the host actually deposits new data.

This fixes the issue on ARM64 where this data was not actually written.
Thus, on a cold boot where memory is zeroed, the correct value of vtom =
0 was read, but during a servicing operation because this memory was not
part of the launch context, the value could be garbage and cause
failures later during initialization.

In the future, we should just have the bootloader parse the information
here, but getting the SLIT and PPTT info reconstructed is a bit
trickier. Tracked by #263

---------

Co-authored-by: Brian Perkins <brian.perkins@microsoft.com>