igvmfilegen: refactor VTL0 image configuration (#271)
Instead of having separate VTL0 and VTL2 fields in the igvmfilegen
configuration, have a single "image" field specifying the actual boot
image to load. Then, for the case of embedding UEFI firmware or a Linux
kernel for OpenHCL to load, have explicit fields within the "openhcl"
type of image.

This has a few advantages:

1. It allows a UEFI firmware to be provided even if OpenHCL is being
loaded into VTL0 (i.e., for the nested virtualization configuration).

2. It acknowledges that the paravisor needs to understand the nested
image type--it's not enough that igvmfilegen knows how to load something
into VTL0, because (except for the UEFI case), the paravisor is still
quite involved in the load process.

3. It is more flexible for the future. With this model, we can
experiment with different, more powerful approaches to loading
firmwares:

* We can ship firmwares uninterpreted, e.g. just embed the UEFI firmware
.fd file directly and do a full load from within OpenHCL.

* Or we can continue to partially load them, but at an offset from where
they eventually need to load.

* We can ship multiple firmwares in a single IGVM file. This would be
most interesting for supporting SeaBIOS and mu_mshv in a single image,
or both the test Linux kernel and mu_mshv.

* Or we can load externally provided firmwares, e.g. via vmbfs or
fw_cfg.

Admittedly, most of these are possible with the existing VTL0/VTL2
split, but they are less natural.