Clean up VTL2 memory handling during boot (#270)
In openvmm: mostly stop treating the required memory directive as
anything special and just do what it's documented to do: validate that
all required memory is present. We can't use it to determine the
required "size" of the paravisor because multiple such directives are
possible. And when the paravisor is loaded into VTL0, we can't tell
which directives are for the paravisor and which are for the UEFI
firmware.

(Arguably, the UEFI firmware directives should just not be included in
the IGVM file and the paravisor should enforce things... but the VMM
still should not be intuiting requirements that are not part of the IGVM
definition.)

In openhcl_boot: don't require a size suggestion from the host if we're
in the mode where VTL2 allocates. This is important because if the host
doesn't want to provide one, then it can't really figure out what to
actually put there (since again, we can't require that there's only one
required memory directive).

(As an aside, I don't see how this case is very different from
non-relocating "host" mode except that we get a hint that it's OK to
change the VTL0 layout. But we seem to treat it very differently. A
puzzle for another day.)