serial: add debugcon device (#284)
Ahh, the joys of open-source!
I can finally work on fun openvmm stuff on the weekend from my personal
PC 😄

This PR adds simple `debugcon` serial device, originally implemented in
bochs, but now widely supported in other VMMs such as QEMU, crosvm,
cloud-hypervisor, etc...

It is a no-frills, relocatable, single IO-port mechanism for a Guest to
write a steam of bytes to the host. It can be instantiated by passing
`--debugcon=port:serialmode` to the CLI (e.g:
`--debugcon=0x402,stderr`).

For context: I am playing around with standing up SeaBIOS support in
OpenVMM (on the weekends, just for kicks), and SeaBIOS writes a lot of
useful debug info to a debugcon at port 0x402. Also, IIUC, OVMF can also
be configured to use a debugcon for early-boot debugging, so this device
might come in handy later as well.