CI: Switch mirror trigger pipeline from a GitHub workflow to an ADO pipeline (#311)
It appears that Azure/pipelines@v1 azure-pipeline-variables is not
working as expected: https://github.com/Azure/pipelines/issues/101. This
change moves triggering of the mirror refresh pipeline to be an Azure
DevOps pipeline instead of a GitHub workflow which should allow us to
pass parameters. This will be needed to mirror the newly-created
release/2411 branch.

---------

Co-authored-by: Ben Hillis <benhill@ntdev.microsoft.com>