virt_mshv_vtl: Add VTL 1 interrupt preemption and VINA support (#248)
When VTL 1 has an interrupt pending, and we're planning on running VTL
0, check to see if we should preempt that and run VTL 1 to deliver the
interrupt instead. Also, sets VINA when VTL 0 has a high enough priority
interrupt and we're running VTL 1. Also resets VINA if auto reset is
asked for.

Note that since we don't have a way for the guest to set its VINA
configuration yet that portion of the code is functionally dead (enabled
will never be true). That's coming soon.