openvmm_hcl: add option to hide isolation from the guest (#330)
Add `OPENHCL_HIDE_ISOLATION=1` command line option to hide the isolation
type from the guest and to disable support for isolation-related
enlightenments such as changing host memory protections on pages.

This allows a VM to run without any shared memory, using normal
architectural and Hyper-V interfaces to interact with devices.

This does not work with the vmbus relay, since the relay relies on the
guest knowing to put the ring buffer in shared memory. Similarly, this
doesn't work with VTL0 assigned PCI devices. Nor does it work with the
net_mana backend, since the MANA driver relies on directly DMAing to
shared memory.

With this, we can boot modern and old guests alike, without
enlightenments, on TDX. SNP is not tested and is unlikely to work
without similar fixups.