vmgs: use `SimpleDisk` instead of `BlockStorage` trait (#351)
The VMGS-specific `BlockStorage` trait represents a poorly maintained
subset of the capabilities of `SimpleDisk`. Switch VMGS to use
`SimpleDisk` and delete the duplicate VHD1 storage backend. Port the GET
VMGS backend to implement `SimpleDisk`.

This will make it straightforward to use different block device backends
for the VMGS file in the future, e.g., to use an NVMe or virtio-blk disk
from within OpenHCL.

This will also be used to fix the GED to wrap an arbitrary disk backend,
rather than its current custom in-memory VMGS implementation.