virt_mshv_vtl: flush deferred actions during servicing (#370)
There may be pending deferred actions (i.e., synic events to signal) on
the VP at time of servicing. Ordinarily, these are flushed the next time
we run the VP, but there is no next time when saving for servicing.
These actions must be flushed to the actual partition state to ensure we
don't drop a signal, which would cause a guest hang.

This fixes a regression from a previous change. Previously, the
`ProcessorRunner` was dropped before a servicing operation, and the
`Drop` implementation would flush deferred actions. Now, the runner
never gets dropped until the partition is torn down. Add a path to
explicitly flush deferred actions before saving.