disk_crypt: add encrypting block device filter (#341)
`disk_crypt` implements confidential (but not authenticated) disks by
encrypting block payload with AES-XTS.

Currently this can be used with openvmm on the host, and not yet with
openhcl. The key can be provided to openvmm via a file on the command
line. E.g.,

```
openvmm --disk crypt:key.bin:file:image.img
```

Here, `key.bin` should contain the key in binary format.