hypestv: add force flag to `kill` command (#393)
Normally, `kill` will power off the VM using `hvc.exe`, which uses the
VMMS WMI API internally. In some cases, the VMMS thinks the VM is not in
a state where it can be powered off even though it can be, e.g. if there
is a pending shutdown request.

Add a `force` flag that uses the HCS API instead (via `hcsdiag.exe`),
which skips the VMMS state machine evaluation and just stops the worker
process.