hypestv/ohcldiag-dev: don't attach to serial ports in server mode (#392)
Hyper-V has a feature where, if a serial port's configured named pipe
server already exists when the VM starts, it will connect as a client to
the named pipe server instead of creating a new pipe server. This makes
it possible to ensure that the pipe connection is established before the
VM starts running, guaranteeing that early boot output is not lost.

Unfortunately, after this initial named pipe connection is terminated,
Hyper-V just gives up--it doesn't try to reconnect to another pipe
server, and it doesn't start a pipe server of its own. There's no way
that I can find to reconnect to the serial port after this.

Given this limitation, don't try to use this special named pipe server
mode configuration. Just poll on a timer to try to catch the early boot
output.

Maybe we can fix Hyper-V at some point.

This fixes a bug in hypestv where, once you disconnect a serial port
(either with `serial 1 off` or by exiting hypestv) that was configured
before the VM started, you can never connect to that serial port again
from any application for the lifetime of the VM.