mesh_rpc: correctly handle unknown methods (#396)
Currently, if a client sends a mesh RPC (TTRPC or gRPC) for method M to
a server that doesn't recognize M (e.g., it's an old build), then the
server will treat that as a channel error. At best, this prevents the
client from getting a clear error message. In practice, it causes the
server to fail.

Fix this by wrapping the server's mesh receiver with a type that handles
the unknown method case by failing the request back to the client with a
clear error message. Also handle the case where the input data cannot be
decoded; although this is not an expected case for our scenarios, it is
certainly not something the server should die over.