disk_backend: refactor `SimpleDisk` (#401)
Clean up the way disks are implemented and used.

Instead of using a single trait object (`SimpleDisk`) everywhere both to
access a block device and to implement it, split these uses in two: a
new struct called `Disk`, and a trait `DiskIo` that merges the old
`SimpleDisk` and `AsyncDisk` traits.

This allows `DiskIo` to focus on being a great trait for implementors,
and `Disk` to focus on being a great type for users.

`DiskIo` no longer needs to be trait object safe, so it can use `async
fn`/`impl Future`. It no longer needs to reference `StackFuture`
anywhere, which makes implementations look a lot more normal.

`Disk` provides improved and consistent `Inspect` output, and it caches
immutable metadata values, allowing users to be less defensive about
caching metadata themselves.