mesh_rpc: client: auto connect and reconnect (#397)
Rather than manage connection and reconnection externally to the
mesh_rpc client, follow the pattern of gRPC clients everywhere and
transparently connect and reconnect when necessary. Apply timeouts and
retries as configured on the client and the RPC.

This simplifies handling for `diag_client` in particular, since it
allows for a single `DiagClient` to be used across multiple reboots of a
VM. This simplifies code in `hypestv`, `openvmm_entry`, `ohcldiag-dev`,
and `petri`.