disk_layered: rework layer attach lifecycle infrastructure (#423)
The `on_attach` model needs to be reworked in order to support
disklayer_sqlite,

Namely: in order to initialize a sqlite disk layer `dbhd` file, one must
know the properties of the lower layer at initialization time, in order
to encode that information in the dbhd's metadata table. Unfortunately,
that information isn't available in `new` - its only available in
`on_attach`. Sticking with the existing `on_attach` model would require
disklayer_sqlite to exist in a "half-initialized" state between
initialization and attachment, thus incurring the need to include
runtime checks in each `LayerIo` method in order to determine whether or
not the backing store has been initialized or not. Not ideal.

Instead, this new model hoists the state transition from initialized to
attached into the type-system itself, with `LayeredDisk` invoking an `fn
attach()` method on a layer type-state which implements `LayerAttach`,
in order to obtain a new type-state of the layer post-attach which
implements the `LayerIo` trait, which can then be used at runtime.

While this does make the LayeredDisk initialization logic a bit more
complicated, it should significantly simplify the disk-layer
implementation of sqlitedisk, and other future disk-layer
implementations.

* * *

As a drive-by, I also cleaned up some of the dependency management
around `disklayer_ram`, making it possible to modify its code without
requiring a complete rebuild of the inverse-dep graph of
`guest_emulation_device` and `storvsp`.