openhcl: make nvme keep alive enabled only via environment arg (#464)
NVMe keep alive requires host changes. Make this experimental feature
opt in only via environment arg.

Additionally revert the get protocol changes. The protocol was specified
incorrectly and we cannot have a flag that must be set to 1 to get the
old behavior. This fixes servicing on older hosts that do not know how
to report this flag.