inspect: Remove the no longer needed Option from Node's sensitivity level (#467)
We first added SensitivityLevel in 1.4. To maintain compatibility with
1.2 we had to add this Option. We are now working on 1.6, and hence
outside the support window for 1.2. So we can remove the Option and
require sensitivity_level to always be present.

I think this is the first time we're testing breaking compatibility in
this way. Should be a good smoke test, and easy to revert if needed.