ci: fixes for issues with unsafe reviewers script (#529)
#503 tried to re-enable the unsafe reviewers script but had it had a few
issues:
1. It was diffing with `origin/main`, which meant that if a PR hadn't
merged main the diff could contain changes that weren't introduced in
the PR.
2. When assigning the review, it was unable to find the
`openvmm-unsafe-approvers` team.

To try and fix those issues this change does the following:
1. It computes the merge base of the PR using `git merge-base` to more
accurately replicate the GitHub diff
2. The team has been changed from `openvmm-unsafe-approvers` to
`@microsoft/openvmm-unsafe-approvers`

I tested this on my fork by pushing a change to my main with unsafe
changes and running the action without merging main on a PR on my fork.
I verified that the unsafe change on main wasn't triggering the unsafe
review on a PR without unsafe changes. I'm unable to test the team
change on my fork because `@microsoft/openvmm-unsafe-approvers` isn't a
collaborator. I've also verified that a PR with unsafe changes still
correctly triggers the reviewer assignment.