mesh_remote: don't bridge initial port with user port (#598)
When a mesh connection is established, an initial port pair is
registered in the leader and follower processes. This port pair is used
to exchange an initial message and is then reused as the user port, by
bridging the initial port and the user port.

This reuse of a port for a different purpose is dangerous, because it
assumes that no extraneous messages are still in a queue somewhere. This
is true with today's `mesh_channel` implementation but is hard to
guarantee in general.

Instead of reusing the initial port via bridging, just send the user
port as part of the initial message. This is simpler to reason about and
is probably roughly as efficient.