pal_async: fix handling of synchronous errors on Windows. (#608)
The `pal_async` crate's handling of overlapped IO on Windows would
always use the `NTSTATUS` value stored in the `OVERLAPPED` structure to
see if the operation completed (the value is not `STATUS_PENDING`).
However, if an IO function completes synchronously, and returned an
error, the value in the `OVERLAPPED` structure does not actually get
updated by the kernel.

When doing some testing using `--kernel-vmnic`, in some cases the value
stored in the `OVERLAPPED` structure would be `STATUS_PENDING` even
though the IO had completed with an error (so the state is
`IoState::None`), leading to a panic on `self.try_complete().unwrap()`.

This change modifies the `pal_async` crate to store the returned error
in case of a synchronous error return, and use that in `try_complete`
instead of checking the `OVERLAPPED` structure. It also fixes the
thread-local executor for overlapped IO to check the result, as it
depended solely on the `OVERLAPPED` structure and would also not
correctly detect this situation.