lints: forbid unsafe op in unsafe fn (#604)
We always want `unsafe {}` blocks to be explicit, even when the unsafe
code is part of an unsafe function. Let's be consistent between `unsafe
{}` inside and outside of `unsafe` functions and set the
`unsafe_op_in_unsafe_fn` lint to "forbid" instead of "warn".

There is no practical difference in checked in code, since we deny
warnings in CI. But I find that for consistency I'd prefer to get errors
instead of warnings for this pattern during local development.

(Frankly `unsafe_op_in_unsafe_fn` seems like it should just be an error
at the language level, not a lint. Embedded developers seem to have
successfully pushed back against that.)