mesh_node: remove 'static+Send bounds from MeshField and MeshPayload (#607)
Mesh values need to be `'static` and `Send` because of implementation
details in the current channel and `LocalNode` implementations. This is
conceptually unnecessary since if we are dealing with a message at the
"mesh" level, we are going to serialize it to protobuf. So, we should
not have to constrain the lifetime or cross-thread safety requirements
of such a value.

Upcoming improvements will make these bounds unnecessary. Remove them
from the `MeshField` and `MeshPayload` traits now in anticipation of
these changes. Update existing code to make these bounds explicit where
required; most are harmless, others some can be removed later as
necessary.