VmBus: cleanly exit when vmbusproxy is in use. (#611)
When the vmbusproxy driver is being used (e.g. using the
`--kernel-vmnic` parameter, only available on Windows), OpenVMM would
either panic or hang during exit, because the thread that issues ioctls
to the proxy driver is never terminated.

This change adds a `CancelContext` to the `proxyintegration` module so
that any outstanding IO can be cancelled, and the thread can exit
cleanly when OpenVMM is terminated.