fuzz_nvme_driver: fix fuzzer-found bugs (#585)
Fix some issues in the `NvmeDriver` fuzzer:

* Sanitize the input provided to .read, .write, and .flush to avoid
contract violations (`NvmeDriver` is an interface interal to `NvmeDisk`,
and it's not useful for the fuzzer to find cases where `NvmeDriver`
enforces the contract).
* Fuzzer can handle when .read, .write, .flush return an Error (bad
input can fail an IO; this is fine).
* Graceful check for some bad device behavior, rathern than just crash.
This will change behavior of the `NvmeDriver` in prod, but if we see a
bad device we will still see key errors in our logs.