RFC: storvsp: new fuzzer (#612)
First commit of a storvsp fuzzer. This begs, borrows, and steals from
other test code to get something that compiles and runs.

Coverage shows that the fuzzer (at this point) does hit much of the core
storvsp packet processing, _except_ for the IO path. That remains a work
in progress.