VmBus: move guest event port creation back to open_channel. (#717)
#662 moved the creation of the guest event port (or rather, enabling it
by setting its parameters) to the `complete_open` method, to simplify
error handling. This caused some issues with the IC channels for some
guests, as those devices can trigger an interrupt during open which is
now lost.

This change moves the creation of the guest event port back to the
`open_channel` method, and makes the necessary changes to deal with
`open_channel` now being able to fail as well.

Fixes #715.