Delete and inline the tempfile_helpers crate (#728)
The one function provided by tempfile_helpers had the footgun that files
would be kept around forever instead of properly cleaned up. It also
forced a specific error handling pattern. Remove it, fix the usage of it
in pal_async to properly clean up test files, and inline it into the
only remaining uses in petri, with better error handling.