nvme_driver: fix overflow computing queue size (#703)
When running the nvme_fuzzer locally I came across an addition_overflow
related panic. Computing the queue size for the I/O queues can cause an
addition overflow related panic if value read from the mqes register is
MAX_U16.

The fix restructures how the io queue size is calculated in way that
eliminates the possibility of addition_overflow. Also added a check to
make sure that mqes_z() does not report an invalid value (< 1) in
accordance with the NVMe spec:

![image](https://github.com/user-attachments/assets/9506ff06-9f5a-4a33-8b9e-70794070bdd6)