virt_mshv_vtl: Share common processing between poll_apic implementations (#630)
This PR pulls out the common ordering and processing needed for
poll_apic across TDX and SNP. While there are similarities between this
code and the poll_apic implementation in mshv, this PR does not change
anything in mshv, as the lapic emulation code there is going to be
deleted soon. This also allows this code to live off of
HardwareIsolatedBacking, making it much cleaner than the first
iteration.

Also cleanup SNP's weird register conversions.