virt_mshv_vtl/mshv_x64: Remove support for emulating APIC (#741)
This operating mode was originally added so that we could get automated
CI test coverage of our APIC emulation without needing to run a full
CVM. However, CVM CI tests are now just around the corner. And, as it
turns out, we accidentally weren't using this in our current tests at
all, and it's now broken. Seeing as we haven't been hit with any APIC
bugs in a very long time, and that proper coverage will be coming soon,
just remove it.

More cleanup is definitely possible here, such as removing support for
certain hypercalls that are no longer needed, or moving trait functions
off of Backing and onto HardwareIsolatedBacking. That can all come as
follow ups.