hvdef: make `HvStatus` type distinct from `HvError` (#754)
`HvError` is supposed to represent a non-success status code, but it's
also used for a possibly-successful status code in some places. Fix this
by adding a new `HvStatus` type and making `HvError` a wrapper around
`NonZeroU16`.

This also gives us a nice niche optimization, so that `HvResult<(),
HvError>` is equivalent to `HvStatus`.