vmm_tests: use libtest-mimic to define and run tests (#762)
Instead of using the built-in test harness, use libtest-mimic to define
tests. This makes tests easier to define:

* We can define tests programmatically at runtime instead of just
statically at compile time; this will make it possible to
reduce/eliminate the use of the domain-specific language in the
`vmm_test` macro to define test constraints (supported VMMs,
architectures, etc.).
* We can define and query artifacts requirements separately from running
the tests, making it easier to integrate with flowey.