TDX TLB: Don't add extended ranges to the list (#779)
TDX supports flushing ranges with the full size of `count` that our
HvGvaSimpleRange supports. It does not support flushing large page
ranges. These two cases are indistinguishable at flush time, as they
have the same binary representation, but we can tell them apart when we
add entries to the list, thanks to the flag parameter in the hypercall.

The original HCL code this was based on was used in multiple locations
and so needed to support storing large page ranges. We do not. Invert
the check for them to make things simpler.

Note that this is the first time this code is meaningfully differing
from the original HCL code.