TDX: Copy vtl-private registers to/from userspace per-vtl (#778)
Introduce a new TdxPrivateRegs struct that is stored per-vtl. This
allows userspace operations to operate on arbitrary VTLs, despite the
kernel only holding one set of registers. Copy these registers in and
out of the shared kernel page on every entry and exit. Remove `pub` from
accessors that directly touch private registers in the kernel and add
new accessors that allow vtl-shared state to still be accessed.

Together with #775 fixes #746.