petri: refactor artifact model (#806)
Currently, test artifact resolution is explicitly done in two phases:
the test requests artifacts into a `TestArtifactRequirements` list. And
then later, the test is handed a `TestArtifacts` bag that can be used to
get the paths of the artifacts that were requested.

The problem with this model is that there are no types linking these two
steps together--it is easy to forget to include artifacts in the
requirements and then see runtime errors later. Worse, the artifacts
have to be computed twice in two different places, making it harder to
extend and refactor the code.

Fix this by making the initial requirements phase return
`ResolvedArtifact` types that can later be passed to the test execution
phase for use as `Path`s, eliminating the `TestArtifacts` bag from the
API. Add explicit `ResolvedArtifact` inputs to various test structures
where the test used to implicitly have an artifact dependency, and
update the test macros to call new methods that fill these in with the
standard values.