nvme_driver: tests: Reducing payload amount to 4MB instead of 4GB for the unit tests (#808)
Reported-By: @mattkur
nvme driver unit tests were taking upwards of 2.5s to execute. Upon
inspection we found that some of the tests were allocating 4GB (4 << 30)
for payload memory. This much memory should not be required since the
only use case for this memory is dma buffers used by the NvmeDriver.

After this change all tests should be running in < 0.1s 

![image](https://github.com/user-attachments/assets/0c27f507-3f87-4617-8051-8e33c5292412)