vmbus_core: fix TargetInfo alignment (#807)
In unit tests, `TargetInfo` can be unaligned, causing a panic when
converting from `&[u8]` to `u64` by reference. Since `TargetInfo`
contained `u8`s and a `u32`, the only guarantee was 4-byte alignment.
`u64` requires 8 byte alignment.

Fix this by converting `TargetInfo` to a `bitfield`, since that is the
semantic reason of this struct anyways (a bit-by-bit representation of
64-bit data type).

```
0x00005555556a213f in vmbus_core::protocol::TargetInfo::as_u64 (self=**0x7ffff768408c**) at vm/devices/vmbus/vmbus_core/src/protocol.rs:320
```

Reported-By: @tjones60
Fix Suggested-By: @jstarks