Create a new hv1_structs crate, move vtl_array into it (#833)
This new crate is intended to be a minimal dependency crate for sharing
useful data structures around, analogous to hvdef but for things we
define ourselves, without requiring crates to depend on all of
hv1_hypercall. Today it holds only VtlArray and Set, but I plan to add
more to it soon. Namely, a ProcessorSet to allow for lazy parsing of the
hypervisor's processor bank & mask structures.

See #835 for my planned first new data structure here