Implement a subset of lxutil (#774)
This is 1/n PRs which will implement lxutil directly in OpenVMM. Once
all PRs are complete, this will remove the OpenVMM dependency on
lxutil.dll. This PR specifically implements all of the functions which
take a `LX_UTIL_FS_CONTEXT` as a parameter and the functions they rely
on.

The new functions introduced by this PR generally follow the same
pattern as the ones exposed by lxutil.dll, and as such minimal changes
are required to existing code. For example, `LxUtilFsChmod` has been
replaced by `fs::chmod`.

Additionally, this PR begins the process of moving the `lxutil` crate to
use `windows-rs` instead of `winapi`. The `winapi` crate has not been
updated for four years and is unofficial. The `windows-rs` crate is
published by Microsoft and is updated with the WinAPI metadata. This
conversion is incomplete, and as such the crate in its current form uses
both `winapi` and `windows-rs`.

This PR was tested with the existing lxutil tests and by sharing a drive
with virtiofs and virtio-9p file systems.

---------

Co-authored-by: John Starks <jostarks@microsoft.com>