nvme_driver: renamed the variable `msix` in IoQueueSavedState to `iv` to match the IoQueue (#845)
Currently the IoQueueSavedState struct is only being used to save the
state of the IoQueue struct. However, the same variable is named
differently between the two structs (msix vs iv). Changing the variable
name in the SavedState struct to match that of IoQueue.
Note: this will not be changing the size of msix. I believe that might
cause some breakage when saving the values.