underhill_mem: Wire up TLB access for flushing and locking (#841)
This creates a new trait, TlbFlushLockAccess, that is used to allow
memory protections management code to access the VP's TLB flush and lock
machinery. It then adds a new parameter of this type to certain methods
on ProtectIsolatedMemory, and adds calls to it where appropriate.

This also adds a small hack to avoid an ongoing layering discussion.
We're unsure of the right approach to fix it, but the rest of the code
should be fine as-is.

Open question: Do we have the right calls everywhere needed? I tried to
match up our code with the legacy HCL's, but our design and flow are
very different, so I'm not 100% confident I got them all correct.

Part 1 of #849