xsync: fix cargo-lock regen (#875)
After improving xsync performance in #857, xsync's internal `cargo
update` command now (silently) fails in CI, causing one of the overlay
deps files to be computed incorrectly. This occurs because in CI, the
registry index is not always populated at the point of xsync, and `cargo
update` cannot download it due to the use of `--offline`.

To preserve most of the performance benefits while fixing the issue, try
`cargo update --offline` first and, if that fails, try it again without
`--offline`. Probably `cargo update --workspace` should be fixed to not
touch the network unless needed.

Also, report the errors from these commands so that this will be caught
next time. Like, come on.