Fix lints for Rust 1.85 (#880)
Mostly cleaning up some more map_ors, but the one big change of note is
temporarily allowing the clippy:precedence lint. Previously this lint
only applied to arithmetic operations, however a change in 1.85 causes
it to apply to bitwise operations as well. This is overly noisy, and
unnecessary. And the clippy maintainers agree, as they've since split
out this change into a separate lint, precedence_bits, that is
allow-by-default. However that split will not be present until 1.86. See
https://github.com/rust-lang/rust-clippy/pull/14115 for details on that.