user_driver: move the page allocator from the NVMe driver code (#872)
Decouples the page allocator from the NVMe subsystem, making it
available for broader use. Currently, the page allocator is tightly
integrated with the NVMe code.
However, with the introduction of the DMA framework, there is a need for
a dedicated bounce buffer allocator. Since the existing page allocator
can serve this purpose, moving it to the user driver allows other
consumers to leverage it for DMA bounce buffer allocation and other use
cases.