vmbus_server: support backpressure for posting messages (#886)
This change adds a mechanism for the `GuestMessagePort` trait to apply
backpressure when VmBus tries to send a message. If a message cannot be
immediately sent, VmBus will queue it to retry later. While messages are
pending, the server will not process incoming messages from the guest to
limit the amount of resources used.

In order to support this properly for reserved channels, including
across save/restore, how sending messages to closed reserved channels is
handled is changed slightly. This means that the reserved channel's
message port is now always destroyed and recreated (to send the closed
response) when closing the channel. This way, the message port doesn't
need to stay open if the message is not sent immediately, and doesn't
need to be reconstructed on save/restore.

Because the openvmm synic implementation currently doesn't implement
backpressure, this code path will not be hit in normal openvmm/openhcl
operation. For this reason, the extra save state field is not a concern
for downgrades.

A test was added for saving/restoring pending messages, and other tests
were updated to simulate a random chance of messages not sending
immediately.