flowey: handle empty action id in binary comparison (#891)
From [this
run](https://github.com/microsoft/openvmm/actions/runs/13446070583/job/37571567182?pr=885),
it looks like it's possible to get an empty string as the action id
which isn't currently being checked for in the binary comparison.

This change adds handling for an empty string by converting it into an
error and continuing to traverse commits in that case. I've also added
some logging for this case.